Crate ibc_types_core_client

IBC client-related types.

Modules

events

Types for ABCI Events that inform relayers about IBC client events.

msgs

Definitions of client-related messages that a relayer submits to a chain.

Structs

ClientId

An IBC client identifier.

ClientType

Type of the client, depending on the specific consensus algorithm.

Height

An IBC height, containing a revision number (epoch) and a revision height (block height).

Enums

Error

A catch-all error type.

HeightParseError

An error while parsing a Height.