ibc_types_core_client
Module msgs

Definitions of client-related messages that a relayer submits to a chain.

Specific implementations of these messages can be found, for instance, in ICS 07 for Tendermint-specific chains. A chain handles these messages in two layers: first with the general ICS 02 client handler, which subsequently calls into the chain-specific (e.g., ICS 07) client handler. See: https://github.com/cosmos/ibc/tree/master/spec/core/ics-002-client-semantics#create.

Structs

MsgCreateClient

A type of message that triggers the creation of a new on-chain (IBC) client.

MsgSubmitMisbehaviour

A type of message that submits client misbehaviour proof.

MsgUpdateClient

A type of message that triggers the update of an on-chain (IBC) client with new headers.

MsgUpgradeClient

A type of message that triggers the upgrade of an on-chain (IBC) client.

Enums

ClientMsg